Add control of operating under promiscuous mode
The default of `tshark` / `dumpcap` (or `tcpdump` for the matter), is to
put the interfaces in promiscuous mode. This can be kind of problematic
because it intermixes "passively observing" with "changing a network
configuration" which might cause side effects.

To give more control over the behavior to the user, this commit
introduces a new environment variable which controls the addition (or
lack of) the `--no-promiscuous-mode`: `PROMISCUOUS_MODE`.

Only if set to "yes", the promiscuous mode is activated (the old default
behavior).