Remove pre-defined capabilities on dumpcap binary
`dumpcap` is the capture engine in the pcap-container. Upstream laid out
guidelines to operate `dumpcap` in a secure fashion with the least
amount of permissions needed:

https://wiki.wireshark.org/CaptureSetup/CapturePrivileges#Limiting_capture_permission_to_only_one_group

In the section "Setting network privileges for dumpcap if your kernel
and file system support file capabilities" upstream envisions to put
capabilities to the `dumpcap` binary itself:

   setcap cap_net_raw,cap_net_admin+eip /usr/sbin/dumpcap

Once the `dumpcap` binary received this treatment, the following
observation can be made when running an container without any special
capabilities:

    / # dumpcap -h
    /bin/sh: dumpcap: Operation not permitted

This can be fixed by running the container with additional capabilities:

    $> docker run -it --rm --cap-add NET_ADMIN alpine:3.22
    / # dumpcap -h
    ...

capabilities(7) documents the purpose of CAP_NET_ADMIN:

    Perform various network-related operations:
    * interface configuration;
    * administration of IP firewall, masquerading, and
    * accounting;
    * modify routing tables;
    * bind to any address for transparent proxying;
    * set type-of-service (TOS);
    * clear driver statistics;
    * set promiscuous mode;
    * enabling multicasting;
    * use setsockopt(2) to set the following socket options:
      SO_DEBUG, SO_MARK, SO_PRIORITY (for a priority outside
      the range 0 to 6), SO_RCVBUFFORCE, and SO_SNDBUFFORCE.

Neither of these use case play a role in the regular use of `dumpcap`
_except_ of promiscuous mode.

AlpineLinux made a decision in 2019 to follow the guidelines of
Wireshark and enforce the capabilities onto the produced binary of
`dumpcap` (Commit: 25292898fde6d5aa8791d9411b904c3a3b408c06, see [2],
[3]). As a consequence, a container must be started with NET_ADMIN
capabilities, indepedendent of the use of promiscuous mode or not.

This commit removes the CAP_NET_ADMIN capabilities on the `dumpcap`
binary and thus give the user of the pcap container the choice of
running it either with appropriate capabilities to operate in promiscous
mode or run it without any special capabilties and still allow to
capture traffic.

CAP_NET_RAW ("cap_net_raw+eip") remains as its required to capture
anything when running as non-root.

[1]: https://www.man7.org/linux/man-pages/man7/capabilities.7.html
[2]: https://gitlab.alpinelinux.org/alpine/aports/-/commit/25292898fde6d5aa8791d9411b904c3a3b408c06
[3]: https://gitlab.alpinelinux.org/alpine/aports/-/blob/25292898fde6d5aa8791d9411b904c3a3b408c06/community/wireshark/APKBUILD#L223-231