Allow to run as non-root
`dumpcap` or rather the wireshark-common package of AlpineLinux creates
a `wireshark` group. Any user within that group is, without any special
other permissions, able to call the `dumpcap` binary and start a packet
capture.

This commit creates the `pcap` user with id 65532 and adds it to the
`wireshark` group. As such, it is possible to run the container with
limitted permissions to capture traffic.